# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BottleIQ API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/imports": {
      "get": {
        "tags": [
          "Imports"
        ],
        "summary": "Jobs",
        "operationId": "jobs_imports_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportJobView"
                  },
                  "title": "Response Jobs Imports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/imports/{job_id}/errors": {
      "get": {
        "tags": [
          "Imports"
        ],
        "summary": "Rejected Rows",
        "operationId": "rejected_rows_imports__job_id__errors_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/imports/{kind}": {
      "post": {
        "tags": [
          "Imports"
        ],
        "summary": "Upload",
        "operationId": "upload_imports__kind__post",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "sales",
                "inventory",
                "purchases"
              ],
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_imports__kind__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/smart-orders": {
      "post": {
        "tags": [
          "Smart Orders"
        ],
        "summary": "Create",
        "operationId": "create_smart_orders_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Smart Orders"
        ],
        "summary": "Orders",
        "operationId": "orders_smart_orders_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderView"
                  },
                  "title": "Response Orders Smart Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/smart-orders/{order_id}": {
      "get": {
        "tags": [
          "Smart Orders"
        ],
        "summary": "Detail",
        "operationId": "detail_smart_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Smart Orders"
        ],
        "summary": "Update",
        "operationId": "update_smart_orders__order_id__patch",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/smart-orders/{order_id}/export": {
      "get": {
        "tags": [
          "Smart Orders"
        ],
        "summary": "Export",
        "operationId": "export_smart_orders__order_id__export_get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_upload_imports__kind__post": {
        "properties": {
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "mapping": {
            "type": "string",
            "title": "Mapping",
            "default": "{}"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "store_id",
          "file"
        ],
        "title": "Body_upload_imports__kind__post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportJobView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "import_type": {
            "type": "string",
            "title": "Import Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "rows_imported": {
            "type": "integer",
            "title": "Rows Imported"
          },
          "rows_rejected": {
            "type": "integer",
            "title": "Rows Rejected"
          },
          "rows_duplicate": {
            "type": "integer",
            "title": "Rows Duplicate"
          },
          "error_summary": {
            "items": {
              "$ref": "#/components/schemas/RejectedRow"
            },
            "type": "array",
            "title": "Error Summary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "import_type",
          "status",
          "row_count",
          "rows_imported",
          "rows_rejected",
          "rows_duplicate",
          "error_summary",
          "created_at"
        ],
        "title": "ImportJobView"
      },
...
      "OrderEdit": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineEdit"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "version",
          "lines"
        ],
        "title": "OrderEdit"
      },
      "OrderInput": {
        "properties": {
          "window": {
            "type": "integer",
            "title": "Window",
            "default": 60
          },
          "target_days": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 7.0,
            "title": "Target Days",
            "default": 21
          },
          "service_level": {
            "type": "number",
            "maximum": 0.999,
            "minimum": 0.8,
            "title": "Service Level",
            "default": 0.95
          },
          "dead_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 30.0,
            "title": "Dead Days",
            "default": 90
          },
          "slow_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 30.0,
            "title": "Slow Days",
            "default": 90
          },
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "vendor_id": {
            "type": "string",
            "title": "Vendor Id"
          }
        },
        "type": "object",
        "required": [
          "store_id",
          "vendor_id"
        ],
        "title": "OrderInput"
      },
      "OrderLineEdit": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "cases": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Cases"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "cases"
        ],
        "title": "OrderLineEdit"
      },
      "OrderLineView": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "cases": {
            "type": "integer",
            "title": "Cases"
          },
          "units": {
            "type": "integer",
            "title": "Units"
          },
          "units_per_case": {
            "type": "integer",
            "title": "Units Per Case"
          },
          "unit_cost": {
            "type": "number",
            "title": "Unit Cost"
          },
          "line_total": {
            "type": "number",
            "title": "Line Total"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "sku",
          "product_name",
          "cases",
          "units",
          "units_per_case",
          "unit_cost",
          "line_total",
          "explanation"
        ],
        "title": "OrderLineView"
      },
      "OrderView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "vendor_id": {
            "type": "string",
            "title": "Vendor Id"
          },
          "vendor_name": {
            "type": "string",
            "title": "Vendor Name"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "estimated_total_cost": {
            "type": "number",
            "title": "Estimated Total Cost"
          },
          "minimum_order_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Order Amount"
          },
          "below_minimum": {
            "type": "boolean",
            "title": "Below Minimum"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineView"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "id",
          "store_id",
          "vendor_id",
          "vendor_name",
          "generated_at",
          "status",
          "version",
          "estimated_total_cost",
          "minimum_order_amount",
          "below_minimum",
          "lines"
        ],
        "title": "OrderView"
      },
...
      "RejectedRow": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "row",
          "message"
        ],
        "title": "RejectedRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}